feat(pricing): replace `double` with `BigDecimal` for monetary values
- Update pricing-related fields in entities (e.g., `LicensePlan`, `Payment`, `Invoice`) to use `BigDecimal` for precision.
- Adjust DTOs (`CreateOrderResponse`, `MarketplaceProductProjection`, etc.) to reflect `BigDecimal` changes.
- Refactor pricing logic in `OrderService` and related entity methods for accurate calculations.
- Update relevant unit tests and assertions to incorporate `BigDecimal` usage.
- Modify database schema to accommodate updated precision and scale for monetary fields.

+++ BE/geolicense/src/main/java/com/alexistdev/geolicense/dto/response/CreateOrderResponse.java
@@ -8,6 +8,7 @@
 
 package com.alexistdev.geolicense.dto.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateOrderResponse(
@@ -15,7 +16,7 @@ public record CreateOrderResponse(
 
         String orderNumber,
 
-        double totalAmount,
+        BigDecimal totalAmount,
 
         String currency,
